requirements.txt
pandas==2.2.2
pandas-gbq==0.22.0
google-cloud-bigquery==3.22.0
